Moved service to workbench layer
